requirements.txt
pandas
plotly
python-dotenv
fsspec
smbprotocol
smbclient
